# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个专业的股票数据抓取与技术指标分析系统。支持 MACD、布林带、神奇九转等多种指标。",
        "title": "牟牛 (MouNiu) 股票分析系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "John Melody",
            "url": "https://github.com/johnmelodyme"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/analysis": {
            "get": {
                "description": "获取数据库中每只股票最新的一条指标计算记录，包括 MACD、布林带、RSI 等",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "指标分析"
                ],
                "summary": "获取所有股票的最新指标分析结果",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.StockIndicator"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/analysis/{symbol}": {
            "get": {
                "description": "根据股票代码获取其最近一次计算的技术指标数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "指标分析"
                ],
                "summary": "获取特定股票的最新指标分析结果",
                "parameters": [
                    {
                        "type": "string",
                        "description": "股票代码 (如 SH600519)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.StockIndicator"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/announcement/all": {
            "get": {
                "description": "分页获取所有上市公司的最新公告信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公告"
                ],
                "summary": "获取所有股票公告",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码 (默认 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量 (默认 20, 最大 100)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/announcement/{symbol}": {
            "get": {
                "description": "根据股票代码模糊查询其相关的公告信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公告"
                ],
                "summary": "获取特定股票的公告",
                "parameters": [
                    {
                        "type": "string",
                        "description": "股票代码 (如 600519)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码 (默认 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量 (默认 20, 最大 100)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/current/{ticker}": {
            "get": {
                "description": "获取指定股票最新的一条价格记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股票数据"
                ],
                "summary": "获取特定股票的当前实时价格",
                "parameters": [
                    {
                        "type": "string",
                        "description": "股票代码 (如 SH600519)",
                        "name": "ticker",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CandleStickData"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/stocks/all": {
            "get": {
                "description": "从数据库中检索所有股票的历史 K 线数据，按时间倒序排列",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股票数据"
                ],
                "summary": "获取所有股票的历史价格数据",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CandleStickData"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/{ticker}": {
            "get": {
                "description": "根据股票代码获取其所有的历史 K 线数据记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "股票数据"
                ],
                "summary": "获取特定股票的历史价格数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "股票代码 (如 SH600519)",
                        "name": "ticker",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CandleStickData"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CandleStickData": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "string"
                },
                "今开盘": {
                    "type": "number"
                },
                "市盈率": {
                    "type": "number"
                },
                "当前股价": {
                    "type": "number"
                },
                "成交量": {
                    "type": "number"
                },
                "成交额": {
                    "type": "number"
                },
                "昨收盘": {
                    "type": "number"
                },
                "更新时间": {
                    "type": "string"
                },
                "最低价": {
                    "type": "number"
                },
                "最高价": {
                    "type": "number"
                },
                "涨跌幅": {
                    "type": "number"
                },
                "涨跌额": {
                    "type": "number"
                },
                "股市值": {
                    "type": "number"
                },
                "股票代码": {
                    "type": "string"
                },
                "股票名称": {
                    "type": "string"
                }
            }
        },
        "model.StockIndicator": {
            "type": "object",
            "properties": {
                "boll_lower": {
                    "type": "number"
                },
                "boll_middle": {
                    "type": "number"
                },
                "boll_upper": {
                    "description": "布林带指标",
                    "type": "number"
                },
                "macd_dea": {
                    "type": "number"
                },
                "macd_dif": {
                    "description": "MACD 指标",
                    "type": "number"
                },
                "macd_hist": {
                    "type": "number"
                },
                "rsi": {
                    "description": "RSI 指标",
                    "type": "number"
                },
                "sma_20": {
                    "description": "移动平均线",
                    "type": "number"
                },
                "stock_code": {
                    "type": "string"
                },
                "stock_name": {
                    "type": "string"
                },
                "td9_count": {
                    "description": "神奇九转指标",
                    "type": "integer"
                },
                "td9_is_buy_setup": {
                    "type": "boolean"
                },
                "td9_is_sell_setup": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    }
}